lift KvArgs out of graphql to be shared with graphql and jsonrpc-alt
