+++ crates/sui-indexer-alt-e2e-tests/src/lib.rs
@@ -36,7 +36,6 @@ use sui_indexer_alt_framework::ingestion::ingestion_client::IngestionClientArgs;
 use sui_indexer_alt_framework::pipeline::CommitterConfig;
 use sui_indexer_alt_framework::postgres::schema::watermarks;
 use sui_indexer_alt_graphql::RpcArgs as GraphQlArgs;
-use sui_indexer_alt_graphql::args::KvArgs as GraphQlKvArgs;
 use sui_indexer_alt_graphql::args::SubscriptionArgs;
 use sui_indexer_alt_graphql::config::RpcConfig as GraphQlConfig;
 use sui_indexer_alt_graphql::start_rpc as start_graphql;
@@ -44,9 +43,9 @@ use sui_indexer_alt_jsonrpc::NodeArgs as JsonRpcNodeArgs;
 use sui_indexer_alt_jsonrpc::RpcArgs as JsonRpcArgs;
 use sui_indexer_alt_jsonrpc::config::RpcConfig as JsonRpcConfig;
 use sui_indexer_alt_jsonrpc::start_rpc as start_jsonrpc;
-use sui_indexer_alt_reader::bigtable_reader::BigtableArgs;
 use sui_indexer_alt_reader::consistent_reader::ConsistentReaderArgs;
 use sui_indexer_alt_reader::fullnode_client::FullnodeArgs;
+use sui_indexer_alt_reader::kv_loader::KvArgs;
 use sui_indexer_alt_reader::system_package_task::SystemPackageTaskArgs;
 use sui_kv_rpc::KvRpcServer;
 use sui_kvstore::BigTableClient;
@@ -417,7 +416,7 @@ impl OffchainCluster {
         let (bigtable_client, bigtable_emulator, archival_service) =
             start_archival(client_args.clone(), kv_rpc_address, registry).await?;
 
-        let graphql_kv_args = GraphQlKvArgs {
+        let kv_args = KvArgs {
             ledger_grpc_url: Some(
                 format!("http://{kv_rpc_address}")
                     .parse()
@@ -428,12 +427,10 @@ impl OffchainCluster {
 
         let jsonrpc = start_jsonrpc(
             Some(database_url.clone()),
-            None,
             DbArgs::default(),
-            BigtableArgs::default(),
+            kv_args.clone(),
             consistent_reader_args.clone(),
             jsonrpc_args,
-            JsonRpcNodeArgs::default(),
             SystemPackageTaskArgs::default(),
             jsonrpc_config,
             registry,
@@ -445,7 +442,7 @@ impl OffchainCluster {
             Some(database_url.clone()),
             fullnode_args,
             DbArgs::default(),
-            graphql_kv_args,
+            kv_args,
             consistent_reader_args,
             graphql_args,
             SystemPackageTaskArgs::default(),
