+++ crates/sui-indexer-alt-e2e-tests/tests/jsonrpc_write_api_tests.rs
@@ -23,6 +23,7 @@ use sui_indexer_alt_jsonrpc::config::RpcConfig;
 use sui_indexer_alt_jsonrpc::start_rpc;
 use sui_indexer_alt_reader::bigtable_reader::BigtableArgs;
 use sui_indexer_alt_reader::consistent_reader::ConsistentReaderArgs;
+use sui_indexer_alt_reader::kv_loader::KvArgs;
 use sui_pg_db::DbArgs;
 use sui_pg_db::temp::TempDb;
 use sui_pg_db::temp::get_available_port;
@@ -92,9 +93,8 @@ impl WriteTestCluster {
 
         let rpc_service = start_rpc(
             Some(database_url),
-            None,
             DbArgs::default(),
-            BigtableArgs::default(),
+            KvArgs::default(),
             ConsistentReaderArgs::default(),
             RpcArgs {
                 rpc_listen_address,
