+++ crates/sui-indexer-alt-graphql/src/args.rs
@@ -4,69 +4,15 @@
 use std::path::PathBuf;
 
 use sui_indexer_alt_metrics::MetricsArgs;
-use sui_indexer_alt_reader::bigtable_reader::BigtableArgs;
 use sui_indexer_alt_reader::consistent_reader::ConsistentReaderArgs;
 use sui_indexer_alt_reader::fullnode_client::FullnodeArgs;
-use sui_indexer_alt_reader::ledger_grpc_reader::LedgerGrpcArgs;
+use sui_indexer_alt_reader::kv_loader::KvArgs;
 use sui_indexer_alt_reader::pg_reader::db::DbArgs;
 use sui_indexer_alt_reader::system_package_task::SystemPackageTaskArgs;
 use tonic::transport::Uri;
 use url::Url;
 
 use crate::RpcArgs;
-
-/// Arguments for configuring KV store access (either Bigtable or Ledger gRPC).
-///
-/// These options are mutually exclusive - only one KV store source can be configured at a time.
-#[derive(clap::Args, Debug, Clone, Default)]
-#[group(required = false)]
-pub struct KvArgs {
-    /// Bigtable instance ID to make KV store requests to.
-    #[arg(long, group = "kv_source")]
-    pub bigtable_instance: Option<String>,
-
-    /// GCP project ID for the BigTable instance (defaults to the token provider's project).
-    #[arg(long)]
-    pub bigtable_project: Option<String>,
-
-    /// App profile ID to use for Bigtable client. If not provided, the default profile will be used.
-    #[arg(long)]
-    pub bigtable_app_profile_id: Option<String>,
-
-    /// Maximum gRPC decoding message size for KV responses, in bytes.
-    ///
-    /// Applies to both Bigtable and Ledger gRPC readers.
-    #[arg(long)]
-    pub kv_max_decoding_message_size: Option<usize>,
-
-    /// gRPC endpoint URL for the ledger service (e.g., archive.mainnet.sui.io)
-    #[arg(long, group = "kv_source")]
-    pub ledger_grpc_url: Option<Uri>,
-
-    /// Time spent waiting for a request to the kv store to complete, in milliseconds.
-    #[arg(long)]
-    pub kv_statement_timeout_ms: Option<u64>,
-}
-
-impl KvArgs {
-    /// Extract BigtableArgs from KvArgs
-    pub fn bigtable_args(&self) -> BigtableArgs {
-        BigtableArgs {
-            bigtable_statement_timeout_ms: self.kv_statement_timeout_ms,
-            bigtable_project: self.bigtable_project.clone(),
-            bigtable_app_profile_id: self.bigtable_app_profile_id.clone(),
-            bigtable_max_decoding_message_size: self.kv_max_decoding_message_size,
-        }
-    }
-
-    /// Extract LedgerGrpcArgs from KvArgs
-    pub fn ledger_grpc_args(&self) -> LedgerGrpcArgs {
-        LedgerGrpcArgs {
-            ledger_grpc_statement_timeout_ms: self.kv_statement_timeout_ms,
-            ledger_grpc_max_decoding_message_size: self.kv_max_decoding_message_size,
-        }
-    }
-}
 
 /// Arguments for configuring GraphQL streaming subscriptions.
 #[derive(clap::Args, Debug, Clone, Default)]
