+++ crates/sui-indexer-alt-jsonrpc/src/context.rs
@@ -7,11 +7,12 @@ use anyhow::Context as _;
 use async_graphql::dataloader::DataLoader;
 use diesel::QueryDsl;
 use prometheus::Registry;
-use sui_indexer_alt_reader::bigtable_reader::BigtableArgs;
 use sui_indexer_alt_reader::bigtable_reader::BigtableReader;
 use sui_indexer_alt_reader::consistent_reader::ConsistentReader;
 use sui_indexer_alt_reader::consistent_reader::ConsistentReaderArgs;
+use sui_indexer_alt_reader::kv_loader::KvArgs;
 use sui_indexer_alt_reader::kv_loader::KvLoader;
+use sui_indexer_alt_reader::ledger_grpc_reader::LedgerGrpcReader;
 use sui_indexer_alt_reader::package_resolver::DbPackageStore;
 use sui_indexer_alt_reader::package_resolver::PackageCache;
 use sui_indexer_alt_reader::pg_reader::PgReader;
@@ -59,15 +60,18 @@ pub(crate) struct Context {
 }
 
 impl Context {
-    /// Set-up access to the stores through all the interfaces available in the context. If
-    /// `bigtable_instance` is set, KV lookups will be sent to it, otherwise they will be sent to
-    /// the `database. If `database_url` is `None`, the interfaces will be set-up but will fail to
+    /// Set-up access to the stores through all the interfaces available in the context.
+    ///
+    /// KV lookups are routed based on `kv_args`: if a Bigtable instance is configured, lookups go
+    /// directly to Bigtable; if a Ledger gRPC URL is configured, lookups go through kv-rpc;
+    /// otherwise they fall back to Postgres.
+    ///
+    /// If `database_url` is `None`, the Postgres-backed interfaces will be set-up but will fail to
     /// accept any connections.
     pub(crate) async fn new(
         database_url: Option<Url>,
-        bigtable_instance: Option<String>,
         db_args: DbArgs,
-        bigtable_args: BigtableArgs,
+        kv_args: KvArgs,
         consistent_reader_args: ConsistentReaderArgs,
         config: RpcConfig,
         metrics: Arc<RpcMetrics>,
@@ -77,16 +81,27 @@ impl Context {
         let pg_reader = PgReader::new(None, database_url, db_args, registry).await?;
         let pg_loader = Arc::new(pg_reader.as_data_loader());
 
-        let kv_loader = if let Some(instance_id) = bigtable_instance {
+        let kv_loader = if let Some(instance_id) = kv_args.bigtable_instance.as_ref() {
             let bigtable_reader = BigtableReader::new(
-                instance_id,
+                instance_id.clone(),
                 "indexer-alt-jsonrpc".to_owned(),
-                bigtable_args,
+                kv_args.bigtable_args(),
                 registry,
             )
             .await?;
 
             KvLoader::new_with_bigtable(Arc::new(bigtable_reader.as_data_loader()))
+        } else if let Some(ledger_grpc_url) = kv_args.ledger_grpc_url.as_ref() {
+            let reader = LedgerGrpcReader::new(
+                ledger_grpc_url.clone(),
+                kv_args.ledger_grpc_args(),
+                Some("jsonrpc_ledger_grpc"),
+                registry,
+            )
+            .await
+            .context("Failed to create Ledger gRPC reader")?;
+
+            KvLoader::new_with_ledger_grpc(Arc::new(reader.as_data_loader()))
         } else {
             KvLoader::new_with_pg(pg_loader.clone())
         };
