+++ crates/sui-indexer-alt-reader/src/kv_loader.rs
@@ -25,17 +25,53 @@ use sui_types::messages_checkpoint::CheckpointSummary;
 use sui_types::object::Object;
 use sui_types::signature::GenericSignature;
 use sui_types::transaction::TransactionData;
+use tonic::transport::Uri;
 
+use crate::bigtable_reader::BigtableArgs;
 use crate::bigtable_reader::BigtableReader;
 use crate::checkpoints::CheckpointKey;
 use crate::error::Error;
 use crate::events::StoredTransactionEvents;
 use crate::events::TransactionEventsKey;
 use crate::ledger_grpc_reader::CheckpointedTransaction;
+use crate::ledger_grpc_reader::LedgerGrpcArgs;
 use crate::ledger_grpc_reader::LedgerGrpcReader;
 use crate::objects::VersionedObjectKey;
 use crate::pg_reader::PgReader;
 use crate::transactions::TransactionKey;
+
+/// Arguments for configuring KV store access (either Bigtable or Ledger gRPC).
+///
+/// These options are mutually exclusive - only one KV store source can be configured at a time.
+#[derive(clap::Args, Debug, Clone, Default)]
+#[group(required = false)]
+pub struct KvArgs {
+    /// Bigtable instance ID to make KV store requests to.
+    #[arg(long, group = "kv_source")]
+    pub bigtable_instance: Option<String>,
+
+    /// GCP project ID for the BigTable instance (defaults to the token provider's project).
+    #[arg(long)]
+    pub bigtable_project: Option<String>,
+
+    /// App profile ID to use for Bigtable client. If not provided, the default profile will be used.
+    #[arg(long)]
+    pub bigtable_app_profile_id: Option<String>,
+
+    /// Maximum gRPC decoding message size for KV responses, in bytes.
+    ///
+    /// Applies to both Bigtable and Ledger gRPC readers.
+    #[arg(long, alias = "bigtable-max-decoding-message-size")]
+    pub kv_max_decoding_message_size: Option<usize>,
+
+    /// gRPC endpoint URL for the ledger service (e.g., archive.mainnet.sui.io)
+    #[arg(long, group = "kv_source")]
+    pub ledger_grpc_url: Option<Uri>,
+
+    /// Time spent waiting for a request to the kv store to complete, in milliseconds.
+    #[arg(long, alias = "bigtable-statement-timeout-ms")]
+    pub kv_statement_timeout_ms: Option<u64>,
+}
 
 /// A loader for point lookups in kv stores backed by either Bigtable, Postgres, or KV gRPC.
 /// Supported lookups:
@@ -90,6 +126,24 @@ pub enum TransactionEventsContents {
 pub enum BalanceChangeContents {
     Grpc(grpc::BalanceChange),
     Native(NativeBalanceChange),
+}
+
+impl KvArgs {
+    pub fn bigtable_args(&self) -> BigtableArgs {
+        BigtableArgs {
+            bigtable_statement_timeout_ms: self.kv_statement_timeout_ms,
+            bigtable_project: self.bigtable_project.clone(),
+            bigtable_app_profile_id: self.bigtable_app_profile_id.clone(),
+            bigtable_max_decoding_message_size: self.kv_max_decoding_message_size,
+        }
+    }
+
+    pub fn ledger_grpc_args(&self) -> LedgerGrpcArgs {
+        LedgerGrpcArgs {
+            ledger_grpc_statement_timeout_ms: self.kv_statement_timeout_ms,
+            ledger_grpc_max_decoding_message_size: self.kv_max_decoding_message_size,
+        }
+    }
 }
 
 impl KvLoader {
